+++ pkg/scalers/aws_cloudwatch_scaler.go
@@ -213,7 +213,7 @@ func (c *awsCloudwatchScaler) GetMetrics(ctx context.Context, metricName string,
 	return append([]external_metrics.ExternalMetricValue{}, metric), nil
 }
 
-func (c *awsCloudwatchScaler) GetMetricSpecForScaling() []v2beta2.MetricSpec {
+func (c *awsCloudwatchScaler) GetMetricSpecForScaling(context.Context) []v2beta2.MetricSpec {
 	targetMetricValue := resource.NewQuantity(int64(c.metadata.targetMetricValue), resource.DecimalSI)
 	externalMetric := &v2beta2.ExternalMetricSource{
 		Metric: v2beta2.MetricIdentifier{
@@ -238,7 +238,7 @@ func (c *awsCloudwatchScaler) IsActive(ctx context.Context) (bool, error) {
 	return val > c.metadata.minMetricValue, nil
 }
 
-func (c *awsCloudwatchScaler) Close() error {
+func (c *awsCloudwatchScaler) Close(context.Context) error {
 	return nil
 }