+++ pkg/scalers/azure_blob_scaler.go
@@ -180,11 +180,11 @@ func (s *azureBlobScaler) IsActive(ctx context.Context) (bool, error) {
 	return length > 0, nil
 }
 
-func (s *azureBlobScaler) Close() error {
+func (s *azureBlobScaler) Close(context.Context) error {
 	return nil
 }
 
-func (s *azureBlobScaler) GetMetricSpecForScaling() []v2beta2.MetricSpec {
+func (s *azureBlobScaler) GetMetricSpecForScaling(context.Context) []v2beta2.MetricSpec {
 	targetBlobCount := resource.NewQuantity(int64(s.metadata.targetBlobCount), resource.DecimalSI)
 	externalMetric := &v2beta2.ExternalMetricSource{
 		Metric: v2beta2.MetricIdentifier{