+++ pkg/scalers/rabbitmq_scaler.go
@@ -336,7 +336,7 @@ func getConnectionAndChannel(host string) (*amqp.Connection, *amqp.Channel, erro
 }
 
 // Close disposes of RabbitMQ connections
-func (s *rabbitMQScaler) Close() error {
+func (s *rabbitMQScaler) Close(context.Context) error {
 	if s.connection != nil {
 		err := s.connection.Close()
 		if err != nil {
@@ -446,7 +446,7 @@ func (s *rabbitMQScaler) getQueueInfoViaHTTP() (*queueInfo, error) {
 }
 
 // GetMetricSpecForScaling returns the MetricSpec for the Horizontal Pod Autoscaler
-func (s *rabbitMQScaler) GetMetricSpecForScaling() []v2beta2.MetricSpec {
+func (s *rabbitMQScaler) GetMetricSpecForScaling(context.Context) []v2beta2.MetricSpec {
 	metricValue := resource.NewQuantity(int64(s.metadata.value), resource.DecimalSI)
 	externalMetric := &v2beta2.ExternalMetricSource{
 		Metric: v2beta2.MetricIdentifier{