+++ pkg/scalers/redis_scaler_test.go
@@ -1,6 +1,7 @@
 package scalers
 
 import (
+	"context"
 	"errors"
 	"testing"
 
@@ -77,14 +78,14 @@ func TestRedisGetMetricSpecForScaling(t *testing.T) {
 			t.Fatal("Could not parse metadata:", err)
 		}
 		closeFn := func() error { return nil }
-		lengthFn := func() (int64, error) { return -1, nil }
+		lengthFn := func(context.Context) (int64, error) { return -1, nil }
 		mockRedisScaler := redisScaler{
 			meta,
 			closeFn,
 			lengthFn,
 		}
 
-		metricSpec := mockRedisScaler.GetMetricSpecForScaling()
+		metricSpec := mockRedisScaler.GetMetricSpecForScaling(context.Background())
 		metricName := metricSpec[0].External.Metric.Name
 		if metricName != testData.name {
 			t.Error("Wrong External metric source name:", metricName)