+++ pkg/scalers/scaler.go
@@ -43,12 +43,12 @@ type Scaler interface {
 
 	// Returns the metrics based on which this scaler determines that the ScaleTarget scales. This is used to construct the HPA spec that is created for
 	// this scaled object. The labels used should match the selectors used in GetMetrics
-	GetMetricSpecForScaling() []v2beta2.MetricSpec
+	GetMetricSpecForScaling(ctx context.Context) []v2beta2.MetricSpec
 
 	IsActive(ctx context.Context) (bool, error)
 
 	// Close any resources that need disposing when scaler is no longer used or destroyed
-	Close() error
+	Close(ctx context.Context) error
 }
 
 // PushScaler interface