+++ pkg/scalers/stan_scaler.go
@@ -115,7 +115,7 @@ func parseStanMetadata(config *ScalerConfig) (stanMetadata, error) {
 func (s *stanScaler) IsActive(ctx context.Context) (bool, error) {
 	monitoringEndpoint := s.getMonitoringEndpoint()
 
-	req, err := http.NewRequest("GET", monitoringEndpoint, nil)
+	req, err := http.NewRequestWithContext(ctx, "GET", monitoringEndpoint, nil)
 	if err != nil {
 		return false, err
 	}
@@ -126,7 +126,7 @@ func (s *stanScaler) IsActive(ctx context.Context) (bool, error) {
 	}
 
 	if resp.StatusCode == 404 {
-		req, err := http.NewRequest("GET", s.getSTANChannelsEndpoint(), nil)
+		req, err := http.NewRequestWithContext(ctx, "GET", s.getSTANChannelsEndpoint(), nil)
 		if err != nil {
 			return false, err
 		}
@@ -196,7 +196,7 @@ func (s *stanScaler) hasPendingMessage() bool {
 	return false
 }
 
-func (s *stanScaler) GetMetricSpecForScaling() []v2beta2.MetricSpec {
+func (s *stanScaler) GetMetricSpecForScaling(context.Context) []v2beta2.MetricSpec {
 	targetMetricValue := resource.NewQuantity(s.metadata.lagThreshold, resource.DecimalSI)
 	metricName := kedautil.NormalizeString(fmt.Sprintf("%s-%s-%s-%s", "stan", s.metadata.queueGroup, s.metadata.durableName, s.metadata.subject))
 	externalMetric := &v2beta2.ExternalMetricSource{
@@ -216,7 +216,7 @@ func (s *stanScaler) GetMetricSpecForScaling() []v2beta2.MetricSpec {
 
 // GetMetrics returns value for a supported metric and an error if there is a problem getting the metric
 func (s *stanScaler) GetMetrics(ctx context.Context, metricName string, metricSelector labels.Selector) ([]external_metrics.ExternalMetricValue, error) {
-	req, err := http.NewRequest("GET", s.getMonitoringEndpoint(), nil)
+	req, err := http.NewRequestWithContext(ctx, "GET", s.getMonitoringEndpoint(), nil)
 	if err != nil {
 		return nil, err
 	}
@@ -244,6 +244,6 @@ func (s *stanScaler) GetMetrics(ctx context.Context, metricName string, metricSe
 }
 
 // Nothing to close here.
-func (s *stanScaler) Close() error {
+func (s *stanScaler) Close(context.Context) error {
 	return nil
 }